refactor: refatora estrutura e componentes relacionados com paginatedTable
