+++ src/features/acf/modules/AcfDashboardPage/modules/PanelSelector/modules/dashboards/modules/PaginatedTable/modules/DataTable/index.tsx
@@ -1,0 +1,47 @@
+import type { GridPaginationModel } from "@mui/x-data-grid";
+import type { Dispatch, SetStateAction, } from "react";
+import { tableDataHook } from "./DataTable.hook";
+import { diabetesColumns } from "./modules/diabetes/modules/columns/columns";
+import type { AcfDashboardType } from "@/features/acf/modules/AcfDashboardPage/types";
+import { Table } from "@impulsogov/design-system";
+import { EmptyTableMessage } from "./modules/EmptyTableMessage";
+
+type TableProps = {
+    acfDashboardType: AcfDashboardType;
+    paginationModel: GridPaginationModel;
+    onPaginationModelChange: Dispatch<SetStateAction<GridPaginationModel>>
+}
+
+export const DataTable = ({
+    acfDashboardType,
+    paginationModel,
+    onPaginationModelChange,
+}: TableProps) => {
+
+    const { data, status, isLoading } = tableDataHook(acfDashboardType, paginationModel);
+
+    if (data && status !== 200 && data.totalRows !== undefined){
+        return <p style={{ textAlign: "center", padding: "20px" }}>
+            Erro ao buscar dados, entre em contato com o suporte.
+        </p>
+    }
+    
+    return (
+        //TODO: Trocar diabetesColumns quando tivermos novas listas
+        <Table
+            columns={diabetesColumns}
+            data={data?.data || []}
+            rowHeight={60}
+            paginationMode="server"
+            // sortingMode="server"
+            rowCount={data?.totalRows || 0}
+            paginationModel={paginationModel}
+            onPaginationModelChange={onPaginationModelChange}
+            // sortModel={DEFAULT_SORTING} //TODO: Mudar para o sorting recebido como prop quando implementarmos sorting
+            // onSortModelChange={(_) => {}} //TODO: Mudar para a função recebida como prop quando implementarmos sorting
+            isLoading={isLoading}
+            slots={{ noRowsOverlay: EmptyTableMessage }}
+            data-testid="list-table"
+        />
+    );
+};
