+++ src/features/acf/modules/AcfDashboardPage/modules/PanelSelector/modules/dashboards/modules/print/buildPrintProps.ts
@@ -5,7 +5,7 @@ import type { FilterItem } from "@/services/lista-nominal/ListaNominal";
 import { PROFILE_ID, type ProfileIdValue } from "@/types/profile";
 import { filtersLabels } from "../filters/filtersLabels";
 import type { DiabetesAcfPrintItem } from "./diabetes/print.model";
-import { diabetesColumns } from "../table/modules/diabetes/modules/columns/columns";
+import { diabetesColumns } from "../PaginatedTable/modules/DataTable/modules/diabetes/modules/columns/columns";
 
 export const buildPrintProps = (
   list: AcfDashboardType,
